improvement: (cr-based) allow an arbitrary number of maven phases
